refactor: calculateDamageWithContextを共通utilsに移動
issue #40で指摘された問題を修正：
- calculateDamageMatrixVaryingAttackとcalculateDamageMatrixVaryingDefenseが
  calculateDamageツールの内部ヘルパーを直接参照していた問題を解決
- calculateDamageWithContextをsrc/utils/に移動して共通化
- 各ツール間の依存関係を適切に整理

Fixes #40
